Issue 4198 - Support Azure Hierarchical workspace storages
