+++ api/build.gradle
@@ -221,6 +221,7 @@ dependencies {
     compile group: "com.azure.resourcemanager", name: "azure-resourcemanager", version: "2.36.0"
     // Azure Storage Blob
     compile group: "com.azure", name: "azure-storage-blob", version: "12.25.0"
+    compile group: "com.azure", name: "azure-storage-file-datalake", version: "12.18.0"
 
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: '2.17.2'
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.17.2'
