+++ api/src/main/java/com/epam/pipeline/controller/vo/region/AzureRegionDTO.java
@@ -41,4 +41,5 @@ public class AzureRegionDTO extends AbstractCloudRegionDTO {
     private String priceOfferId;
     private Boolean enterpriseAgreements;
     private String managedIdentity;
+    private boolean hierarchicalNsStorage;
 }
