+++ api/src/main/java/com/epam/pipeline/dao/region/AzureRegionDaoHelper.java
@@ -54,6 +54,7 @@ class AzureRegionDaoHelper extends AbstractCloudRegionDaoHelper<AzureRegion, Azu
         params.addValue(CloudRegionParameters.METER_REGION_NAME.name(), region.getMeterRegionName());
         params.addValue(CloudRegionParameters.AZURE_API_URL.name(), region.getAzureApiUrl());
         params.addValue(CloudRegionParameters.MANAGED_IDENTITY.name(), region.getManagedIdentity());
+        params.addValue(CloudRegionParameters.HIERARCHICAL_NS_STORAGE.name(), region.isHierarchicalNsStorage());
         params.addValue(CloudRegionParameters.PRICE_OFFER_ID.name(), region.getPriceOfferId());
         params.addValue(CloudRegionParameters.ENTERPRISE_AGREEMENTS.name(), region.isEnterpriseAgreements());
         return params;
@@ -74,6 +75,7 @@ class AzureRegionDaoHelper extends AbstractCloudRegionDaoHelper<AzureRegion, Azu
         region.setMeterRegionName(rs.getString(CloudRegionParameters.METER_REGION_NAME.name()));
         region.setAzureApiUrl(rs.getString(CloudRegionParameters.AZURE_API_URL.name()));
         region.setManagedIdentity(rs.getString(CloudRegionParameters.MANAGED_IDENTITY.name()));
+        region.setHierarchicalNsStorage(rs.getBoolean(CloudRegionParameters.HIERARCHICAL_NS_STORAGE.name()));
         region.setPriceOfferId(rs.getString(CloudRegionParameters.PRICE_OFFER_ID.name()));
         region.setEnterpriseAgreements(rs.getBoolean(CloudRegionParameters.ENTERPRISE_AGREEMENTS.name()));
         return region;
