+++ api/src/main/java/com/epam/pipeline/dao/region/CloudRegionParameters.java
@@ -69,5 +69,6 @@ enum CloudRegionParameters {
     USER_PASSWORD,
     CLUSTER_INCLUDE,
     CLUSTER_STATE_PROPERTIES,
-    MANAGED_IDENTITY
+    MANAGED_IDENTITY,
+    HIERARCHICAL_NS_STORAGE
 }
