+++ api/src/main/resources/dao/cloud-region-dao.xml
@@ -50,6 +50,7 @@
                         meter_region_name,
                         azure_api_url,
                         managed_identity,
+                        hierarchical_ns_storage,
                         price_offer_id,
                         enterprise_agreements,
                         project,
@@ -99,6 +100,7 @@
                         :METER_REGION_NAME,
                         :AZURE_API_URL,
                         :MANAGED_IDENTITY,
+                        :HIERARCHICAL_NS_STORAGE,
                         :PRICE_OFFER_ID,
                         :ENTERPRISE_AGREEMENTS,
                         :PROJECT,
@@ -152,6 +154,7 @@
                         r.meter_region_name,
                         r.azure_api_url,
                         r.managed_identity,
+                        r.hierarchical_ns_storage,
                         r.price_offer_id,
                         r.enterprise_agreements,
                         r.project,
@@ -216,6 +219,7 @@
                         r.meter_region_name,
                         r.azure_api_url,
                         r.managed_identity,
+                        r.hierarchical_ns_storage,
                         r.price_offer_id,
                         r.enterprise_agreements,
                         r.project,
@@ -280,6 +284,7 @@
                         r.meter_region_name,
                         r.azure_api_url,
                         r.managed_identity,
+                        r.hierarchical_ns_storage,
                         r.price_offer_id,
                         r.enterprise_agreements,
                         r.project,
@@ -344,6 +349,7 @@
                         r.meter_region_name,
                         r.azure_api_url,
                         r.managed_identity,
+                        r.hierarchical_ns_storage,
                         r.price_offer_id,
                         r.enterprise_agreements,
                         r.project,
@@ -408,6 +414,7 @@
                         r.meter_region_name,
                         r.azure_api_url,
                         r.managed_identity,
+                        r.hierarchical_ns_storage,
                         r.price_offer_id,
                         r.enterprise_agreements,
                         r.project,
@@ -472,6 +479,7 @@
                         meter_region_name = :METER_REGION_NAME,
                         azure_api_url = :AZURE_API_URL,
                         managed_identity = :MANAGED_IDENTITY,
+                        hierarchical_ns_storage = :HIERARCHICAL_NS_STORAGE,
                         price_offer_id = :PRICE_OFFER_ID,
                         enterprise_agreements = :ENTERPRISE_AGREEMENTS,
                         project = :PROJECT,
@@ -540,6 +548,7 @@
                         r.meter_region_name,
                         r.azure_api_url,
                         r.managed_identity,
+                        r.hierarchical_ns_storage,
                         r.price_offer_id,
                         r.enterprise_agreements,
                         r.project,
