+++ api/src/main/resources/db/migration/v2025.12.02_12.00__azure_hierarchical_ns_storage.sql
@@ -1,0 +1,1 @@
+ALTER TABLE pipeline.cloud_region ADD COLUMN hierarchical_ns_storage BOOLEAN NULL;
