+++ core/src/main/java/com/epam/pipeline/entity/region/AzureRegion.java
@@ -52,4 +52,5 @@ public class AzureRegion extends AbstractCloudRegion {
     private String priceOfferId;
     private boolean enterpriseAgreements;
     private String managedIdentity;
+    private boolean hierarchicalNsStorage;
 }
